microservices/user_service/requirements.txt
Django==3.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
Pillow==10.1.0
pytz==2024.1
sqlparse==0.4.4
asgiref==3.7.2 